updated configuration
