+++ src/app.module.ts
@@ -7,33 +7,11 @@ import { ConfigModule } from '@nestjs/config';
 import configuration from './config/configuration';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Connection } from 'typeorm';
-import { User } from './orm/user/user.entity';
-import { Account } from './orm/account/account.entity';
-import { Company } from './orm/company/company.entity';
-import { Role } from './orm/role/role.entity';
-import { Membership } from './orm/membership/membership.entity';
-
-const conf = configuration();
+import ormconfig from 'ormconfig';
 
 @Module({
   imports: [
-    TypeOrmModule.forRoot({
-      name: 'default',
-      type: 'postgres',
-      host: conf.db.host,
-      port: conf.db.port,
-      username: conf.db.username,
-      password: conf.db.password,
-      database: conf.db.name,
-      entities: [User, Account, Company, Role, Membership],
-      migrationsTableName: 'custom_migration_table',
-      migrations: ['migration/*.ts'],
-      cli: {
-        migrationsDir: 'migration',
-      },
-      // synchronize: true,
-      verboseRetryLog: true,
-    }),
+    TypeOrmModule.forRoot(ormconfig),
     AuthModule,
     UsersModule,
     ConfigModule.forRoot({
